+++ docs/about/history.rst
@@ -53,6 +53,7 @@ Fixed
    -  Fix bug where wind direction could not be set per ShapeGerstner component.
    -  Fix compilation errors when only Unity's new *Input System* backend is available.
    -  Fix null exceptions in validation when *OceanRenderer* is not present.
+   -  Fix incorrect validation showing in prefab mode.
 
    .. only:: hdrp